remove: auth 관련 파일 정리
액션 함수 auth..ts 파일로 정리
